+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/settings/ProjectSettings.kt
@@ -33,6 +33,9 @@ class ProjectSettings : PersistentStateComponent<ProjectSettings?> {
     @Attribute
     var isProjectSpecificLLMClient: Boolean = false
 
+    @Transient
+    var splitButtonActionSelectedLLMClientId: String? = null
+
     @Property
     var activePrompt = DefaultPrompts.BASIC.prompt
         get() = getActivePrompt(field)
@@ -59,6 +62,13 @@ class ProjectSettings : PersistentStateComponent<ProjectSettings?> {
         } else {
             AppSettings2.instance.getActiveLLMClientConfiguration()
         }
+    }
+
+    fun getSplitButtonActionSelectedOrActiveLLMClient(): LLMClientConfiguration? {
+        // First try to get the session client, then fall back to the active client
+        return splitButtonActionSelectedLLMClientId?.let {
+            AppSettings2.instance.getActiveLLMClientConfiguration(it)
+        } ?: getActiveLLMClientConfiguration()
     }
 
     private fun getActivePrompt(activePrompt: Prompt): Prompt {