+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [2.16.0] - 2025-09-04
+
 ### Added
 
 - Support AWS SDK default credential locations for Bedrock Integration (#368).
@@ -420,7 +422,8 @@
 - Settings for locale and OpenAI token.
 - Create commit message only for selected files.
 
-[Unreleased]: https://github.com/Blarc/ai-commits-intellij-plugin/compare/v2.15.0...HEAD
+[Unreleased]: https://github.com/Blarc/ai-commits-intellij-plugin/compare/v2.16.0...HEAD
+[2.16.0]: https://github.com/Blarc/ai-commits-intellij-plugin/compare/v2.15.0...v2.16.0
 [2.15.0]: https://github.com/Blarc/ai-commits-intellij-plugin/compare/v2.14.1...v2.15.0
 [2.14.1]: https://github.com/Blarc/ai-commits-intellij-plugin/compare/v2.14.0...v2.14.1
 [2.14.0]: https://github.com/Blarc/ai-commits-intellij-plugin/compare/v2.13.0...v2.14.0